quick fix

+++ src/server/prototypes/player/delete-character.ts
@@ -4,11 +4,19 @@ import { container } from "@shared/ioc-container";
 
 const prisma = container.get(PrismaClient);
 
+declare module "alt-server" {
+  export interface Player {
+    deleteCharacter(characterId: string): Promise<void>;
+  }
+}
+
 Player.prototype.deleteCharacter = async function (characterId: string) {
+  if (!this.store.isLoggedIn) return;
+
   await prisma.character.deleteMany({
     where: {
       id: characterId,
-      userId: this.user!.id,
+      userId: this.store.user.id,
     },
   });
 };
